generalizing NLP model support a bit more
